1. перенесла свои тесты в FooterRestApiLinkTest.java 2. удалила свой класс RestApiLinkTest.java 3. добавила 2 теста в FooterVersionMenuTest
