+++ src/test/java/school/redrover/FooterRestApiLinkTest.java
@@ -47,4 +47,75 @@ public class FooterRestApiLinkTest extends BaseTest {
         Assert.assertEquals(restApiLink.getCssValue("cursor"), "pointer");
         Assert.assertNotEquals(beforeBackground, afterBackground);
     }
+
+    @Test (dependsOnMethods = "testRestApiLinkOpensInSameTab")
+    public void testRestApiLinkIsHiddenOnApiPage() {
+
+        JavascriptExecutor js = (JavascriptExecutor) getDriver();
+
+        js.executeScript("window.scrollTo(0, document.body.scrollHeight);");
+
+        getWait10().until(
+                ExpectedConditions.elementToBeClickable(By.xpath("//footer//a[contains(text(),'REST API')]"))
+        ).click();
+
+        js.executeScript("window.scrollTo(0, document.body.scrollHeight);");
+
+        boolean isRestApiLinkPresentInFooter = !getDriver().findElements(
+                By.xpath("//footer//a[contains(text(),'REST API')]")
+        ).isEmpty();
+
+        Assert.assertFalse(isRestApiLinkPresentInFooter,
+                "Ссылка 'REST API' не должна отображаться в футере на странице REST API (сама на себя)");
+    }
+
+    @Test
+    public void testRestApiLinkOpensInSameTab() {
+        JavascriptExecutor js = (JavascriptExecutor) getDriver();
+
+        js.executeScript("window.scrollTo(0, document.body.scrollHeight);");
+
+        WebElement restApiLink = getWait10().until(
+                ExpectedConditions.elementToBeClickable(By.xpath("//footer//a[contains(text(),'REST API')]"))
+        );
+
+        String originalWindow = getDriver().getWindowHandle();
+        restApiLink.click();
+
+        Assert.assertEquals(getDriver().getWindowHandle(), originalWindow,
+                "Фокус переключился на другое окно");
+
+    }
+
+    @Test
+    public void testRestApiLinkHoverEffect() {
+        JavascriptExecutor js = (JavascriptExecutor) getDriver();
+
+        js.executeScript("window.scrollTo(0, document.body.scrollHeight);");
+
+        WebElement restApiLink = getWait10().until(
+                ExpectedConditions.visibilityOfElementLocated(By.xpath("//footer//a[contains(text(),'REST API')]"))
+        );
+
+        String cursor = restApiLink.getCssValue("cursor");
+        Assert.assertEquals(cursor, "pointer", "У ссылки должен быть курсор pointer при наведении");
+    }
+
+    @Test (dependsOnMethods = "testRestApiLinkOpensInSameTab")
+    public void testReturnWithBackButton() {
+        JavascriptExecutor js = (JavascriptExecutor) getDriver();
+
+        js.executeScript("window.scrollTo(0, document.body.scrollHeight);");
+
+        getWait10().until(
+                ExpectedConditions.elementToBeClickable(By.xpath("//footer//a[contains(text(),'REST API')]"))
+        ).click();
+
+        getDriver().navigate().back();
+
+        boolean isDashboardVisible = getWait10().until(
+                ExpectedConditions.visibilityOfElementLocated(By.xpath("//a[contains(@href,'/view/')]"))
+        ).isDisplayed();
+        Assert.assertTrue(isDashboardVisible, "Элементы Dashboard не отображаются. Возможно, пользователь разлогинен.");
+    }
 }
